refactor: dialogコンポーネントの整理
- DatasetSettingDaialog のタイポを DatasetSettingDialog に修正
- JdbcTableSelectorDialog を ResourceFileDialog を使うよう統一（raw <dialog open> から showModal() パターンへ）
- TemplateEditDialog を TemplateEditButton.tsx から独立ファイルに抽出
- components/dialog/index.ts バレルファイルを追加し、各ダイアログのインポートを統一
