+++ tauri/src/components/dialog/index.ts
@@ -1,0 +1,14 @@
+export {
+	SettingDialog,
+	Fieldset,
+	Check,
+	Text,
+	Select,
+	Arrays,
+	ArraysText,
+	KeyValues,
+	KeyValueText,
+} from "./SettingDialog";
+export type { SettingDialogProps } from "./SettingDialog";
+export { default as ResourceFileDialog } from "./ResourceFileDialog";
+export { default as SettingTable } from "./SettingTable";
